Change order of match for GameResults to handle IllegalMove first
Previously it doesn't check IllegalMove first, and instead try to move
the piece first and then check status, if illegal move then rollback.
This is ok for the most part but since we put turn switcher in Continue
it means if illegal moves happen then it will switch to CPU turn.